requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
decorator==4.4.2
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
lxml==5.3.1
MarkupSafe==3.0.2
moviepy==1.0.3
narwhals==1.33.0
networkx==3.4.2
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==10.4.0
proglog==0.1.11
protobuf==5.29.4
pyarrow==19.0.1
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
scikit-image==0.25.2
scipy==1.15.2
six==1.17.0
smmap==5.0.2
streamlit==1.44.1
tenacity==9.1.2
tifffile==2025.3.30
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
watchdog==6.0.0
